pyproject.toml
[tool.poetry]
name = "transitweb"
version = "0.1.0"
description = ""
authors = ["Tyler Gannon <tgannon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.dev.dependencies]
typer = { version = "^0.7.0", extras = ["all"] }
ansible = "^7.3.0"
ansible-lint = { version = "^6.14.0", markers = "platform_system != 'Windows'" }
black = "^23.1.0"
jinja2 = "^3.1.2"
faker = "^18.4.0"
python-lorem = "^1.2.0"
ptpython = "^3.0.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
